add suspense to search page and made server component
